Initial commit of state management overhaul
The major premise of this change is to totally remove support for an
entity being able to have multiple states at the same time. While it is
not immediately apparent, there are inherent limitations for state
machines that cannot be solved while maintaining support for multiple
states at a time.